working on the base classes

+++ router/planningposition.py
@@ -1,12 +1,12 @@
 """ Top planning position for incomes, expenses assets and credits """
-from fastapi import FastAPI
+
 from pydantic import BaseModel
 from typing import Optional
 
 class Planningposition(BaseModel):
     name: str
     person_id: int
-    initialValue: Optional[int] = 0
+    initialValue: Optional[float] = 0
